Fixed empty values
If you use empty values in your env files, e.g. `DB_PORT=` and cast it in call into int or float, so you will get an error:
```
ValueError: invalid literal for int() with base 10: ''
```

I just fixed that for me with this universal solution. It works not only on int or float